requirements/prod.txt
# django
Django>=4.0.5
djangorestframework>=3.13.0
django-grappelli>=3.0.3
django-filter==22.1
django-cleanup==6.0.0
django-storages>=1.12.0
django-imagekit==4.1.0
gunicorn==20.1.0

# security and auth
django-cors-headers==3.13.0
djangorestframework-simplejwt
djoser==2.1.0

# database
psycopg2-binary>=2.9.0

# other
pyyaml==6.0
pillow==9.1.1







